# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Item Test API",
        "description": "API untuk mengelola data dokter phc",
        "version": "1.0"
    },
    "paths": {
...
        "/api/labtransdetail": {
            "get": {
                "tags": [
                    "Lab Detail"
                ],
                "summary": "Get all Lab Trans with a single search parameter",
                "operationId": "e9cf1d304f246ab4293fdbcb13063518",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Search by exact ID (integer match)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search by text query (LIKE match), filtering for 'urine' if provided",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered list of Lab Trans",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LabTransDetail"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Lab Detail"
                ],
                "summary": "Create a new LabTransDetail",
                "operationId": "74562684ef51f3ea04f560a58eb03a9a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "LabTransID": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "ItemTestID": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "ResultValue": {
                                        "type": "string",
                                        "example": "Positive"
                                    },
                                    "Unit": {
                                        "type": "string",
                                        "example": "mg/dL"
                                    },
                                    "ReferenceValue": {
                                        "type": "string",
                                        "example": "70-110"
                                    },
                                    "ResultNotes": {
                                        "type": "string",
                                        "example": "Normal Range"
                                    },
                                    "CreateBy": {
                                        "type": "string",
                                        "example": "Admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data berhasil disimpan"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/LabTransDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/labtransdetail/{ID}": {
            "put": {
                "tags": [
                    "Lab Detail"
                ],
                "summary": "Update a specific LabTransDetail",
                "operationId": "66727ccbf34a96a331899ca86a6f5cfd",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "LabTransDetail ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ItemTestID": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "LabTransID": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "ResultValue": {
                                        "type": "string",
                                        "example": "Positive"
                                    },
                                    "Unit": {
                                        "type": "string",
                                        "example": "mg/dL"
                                    },
                                    "ReferenceValue": {
                                        "type": "string",
                                        "example": "70-110"
                                    },
                                    "ResultNotes": {
                                        "type": "string",
                                        "example": "Normal Range"
                                    },
                                    "LastModifiedBy": {
                                        "type": "string",
                                        "example": "Admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/LabTransDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LabTransDetail": {
                "required": [
                    "LabTransID",
                    "ItemTestID",
                    "ResultValue",
                    "Unit"
                ],
                "properties": {
                    "LabTransID": {
                        "description": "Lab Transaction ID",
                        "type": "integer"
                    },
                    "ItemTestID": {
                        "description": "Item Test ID",
                        "type": "integer"
                    },
                    "ResultValue": {
                        "description": "Result Value",
                        "type": "string"
                    },
                    "Unit": {
                        "description": "Unit of the result",
                        "type": "string"
                    },
                    "ReferenceValue": {
                        "description": "Reference Value",
                        "type": "string"
                    },
                    "ResultNotes": {
                        "description": "Result Notes",
                        "type": "string"
                    },
                    "CreateDate": {
                        "description": "Date when the record was created",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "CreateBy": {
                        "description": "Name of the user who created the record",
                        "type": "string"
                    },
                    "LastModifiedDate": {
                        "description": "Date when the record was last modified",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "LastModifiedBy": {
                        "description": "Name of the user who last modified the record",
                        "type": "string"
                    },
                    "gcrecord": {
                        "description": "Indicator for soft delete",
                        "type": "boolean"
                    },
                    "LabNumber": {
                        "description": "Lab number identifier",
                        "type": "string",
                        "maxLength": 20
                    },
                    "LabTest": {
                        "description": "Test type for the lab transaction",
                        "type": "string",
                        "maxLength": 20
                    },
                    "TransDate": {
                        "description": "Transaction date of the lab test",
                        "type": "string",
                        "format": "date"
                    },
                    "DoctorReferral": {
                        "description": "Doctor who referred the lab test",
                        "type": "string",
                        "maxLength": 50
                    },
                    "PatientID": {
                        "description": "ID of the patient associated with the lab test",
                        "type": "integer"
                    },
                    "Age": {
                        "description": "Patient's age during the lab test",
                        "type": "string",
                        "maxLength": 10
                    },
                    "Anamnesa": {
                        "description": "Patient's medical history description",
                        "type": "string",
                        "maxLength": 250
                    },
                    "BB": {
                        "description": "Patient's body weight",
                        "type": "string",
                        "maxLength": 10
                    },
                    "TB": {
                        "description": "Patient's height",
                        "type": "string",
                        "maxLength": 10
                    },
                    "LP": {
                        "description": "Patient's waist circumference",
                        "type": "string",
                        "maxLength": 10
                    },
                    "TD": {
                        "description": "Patient's blood pressure",
                        "type": "string",
                        "maxLength": 10
                    },
                    "BMI": {
                        "description": "Patient's Body Mass Index",
                        "type": "string",
                        "maxLength": 10
                    },
                    "FinalStatement": {
                        "description": "Doctor's final statement about the test",
                        "type": "string",
                        "maxLength": 250
                    },
                    "FinalResult": {
                        "description": "Final result of the lab test",
                        "type": "string",
                        "maxLength": 20
                    },
                    "Status": {
                        "description": "Status of the lab test",
                        "type": "string",
                        "maxLength": 20
                    }
                },
                "type": "object"
            },
...
        }
    }
}